netstack: performance fixes as suggested by @hbhasker
https://github.com/rootless-containers/rootlesskit/pull/101
a) NewForwarder is being called with a window of 30k which
means its advertising a Window Scale of 1 resulting in every
packet causing a ZeroWindow event/update.
b) RXChecksumOffload should be set to true as when reading packets
from tap there is no need to do checksum verification. Ideally we
should not need to do TXChecksumming either but enabling that causes
linux to drop TCP packets w/o valid checksums. Also our
calculateChecksum code is slow and could use with some loop unrolling
to make it much faster.
c) slirpnetstack is enabling sniffer on the link endpoint which means
its trying to log every packet. This will slow down things dramatically.
This should be made configurable and only enabled when debugging issues.
d) ModerateRecvBuf to true won't really work as gonet API is not
configured to invoke the endpoint.ModerateRecvBuf after Read(). The
way auto-tuning works is gvisor invokes the API after it copies bytes
to user-space. Right now setting it to true will not do anything unless
the API is called. That said there is no real reason to use auto-tuning
since its all onhost connection and a buffer of a couple of MB is more
than enough to hit 10Gbits/s.