Add proxy-protocol support to local forwarding code
Proxy-protocol (PP) and SPP are techniques to preserve real client
IP when using simple TCP or UDP proxies. The proxy adds a header and
then slirpnetstack is able to unpack it, and spoof the source IP internally
to a tunnel. Of course proxying connections like 10.x or 127.x is not
permitted.