Start a DHCP server, make QEMU IPv4 guests happy
I didn't include the DHCPv6 server yet. QEMU implementation seemed
partial, and in general a Linux guest doesn't seem to query it at all,
as it relies on NDP instead (which I am working on).

Currently, there is no option to specify the range (Host[:3].15-100)
or various other DHCP options.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>