Add --disable-host-networks flag, to disable connections to local networks.
When doing routing, add an option to slightly secure the traffic of
potentially untrusted guest - prevent it from connecting to networks
we have direct knowledge of. Specifically - prevent routing to networks
that are in "main" and "local" linux routing tables. Refresh the view
every 30 seconds, in case the routing info changes.

This is a rudimentary technique, but should be sufficient for most
basic uses.