misc: properly support sending packets from 0.0.0.0 using non-default ip
This is kindof a broken BSD sockets API. It turns out that receiving
packet is easy - you can get client IP and live long. The problem is
that when binding to 0.0.0.0 the server/local IP can be anything.
It's surprisingly hard to extract this laddr and then use it as source
when transmitting response.