On firewall block capturing TCP connection, always break it with RST
In past we tried to convey something with either sending or not sending
RST on firewall block. This was misguided and only caused the guest
to hang for longer on firewall block. In future maybe set the reason
for block in ICMP or RST message, but definitely making guest hang
if we _know_ the connection is not allowed was a bad tactic.