major: Remove --disable-routing and --disable-host-networks options, replace them with --enable-routing and --enable-host
The reason for this is security. In the context of slirpnetstack we mostly
want to isolate the guest - this is the whole point of the project.
So let's change the defaults to default deny.

Now --enable-host and --enable-routing are non-overlapping, which I think
makes sense. so: without --enable-host routing to local addresses will fail
without --enable-routing routing to internet will fail